refactor: reorganize authentication ports and update imports
- Moved `AuthenticationService` trait to a new `authentication` module for better structure and clarity.
- Updated import paths across various files to reflect the new module organization.
- Cleaned up unused code in the `ports.rs` file, enhancing maintainability.
