+++ api/src/lib/application/http/server/http_server.rs
@@ -4,8 +4,8 @@ use crate::application::http::realm::router::realm_routes;
 use crate::application::http::server::app_state::AppState;
 use crate::application::http::server::openapi::ApiDoc;
 use crate::application::http::user::router::user_routes;
-use crate::domain::authentication::ports::AuthenticationService;
 use crate::domain::authentication::ports::auth_session::AuthSessionService;
+use crate::domain::authentication::ports::authentication::AuthenticationService;
 use crate::domain::client::ports::ClientService;
 use crate::domain::credential::ports::CredentialService;
 use crate::domain::realm::ports::RealmService;
